+++ util/crystalizer.py
@@ -2,11 +2,10 @@
 
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.interpolate import interp1d
 
 f1, f2 = 50.0, 60.0  # frequencias
 
-t1, t2 = 0.0, 100.0
+t1, t2 = 0.0, 200.0
 
 # frequencia angular
 omega1, omega2 = 2.0 * np.pi * f1, 2.0 * np.pi * f2
@@ -24,39 +23,36 @@ def y(t): return y1(t) + y2(t)
 t = np.linspace(t1, t2, 1000)
 original = y(t)
 
-# interp_f = interp1d(t, original, kind='cubic')
-# t = np.linspace(t1, t2, 200)
-# original = interp_f(t)
-
 ffmpeg = np.copy(original)
 other = np.copy(original)
 
-intensity = 10.0
+intensity = 10.0 * 10
 last_v_ffmpeg = ffmpeg[0]
-last_v_other = other[0]
 
 for n in range(ffmpeg.size):
-    # ffmpeg method
-
     v = ffmpeg[n]
 
     ffmpeg[n] = v + (v - last_v_ffmpeg) * intensity
 
     last_v_ffmpeg = v
 
-    # other
 
+last_v_other = other[-1]
+for n in range(other.size - 1, 0, -1):
     v = other[n]
-
-    other[n] = last_v_other + (v - last_v_other) * intensity
-
+    other[n] = v + (v - last_v_other) * intensity
     last_v_other = v
+
+other = 0.5 * (other + ffmpeg)
+
+other[0] = other[1] + (other[1] - other[2])
+other[-1] = other[-2] + (other[-2] - other[-3])
 
 fig = plt.figure()
 
 plt.plot(t, original, 'bo-', markersize=4, label='original')
-plt.plot(t, ffmpeg, 'ro-', markersize=4, label='ffmpeg')
-plt.plot(t, other, 'go-', markersize=4, label='other')
+# plt.plot(t, ffmpeg, 'go-', markersize=4, label='ffmpeg')
+plt.plot(t, other, 'ro-', markersize=4, label='other')
 
 fig.legend()